scraper/requirements.txt
pymongo
pandas
requests
beautifulsoup4
html5lib
gitpython